storage: move test utiltiies out of test files
