+++ staging/src/k8s.io/apiserver/pkg/storage/testing/utils.go
@@ -19,9 +19,16 @@ package testing
 import (
 	"context"
 	"path"
+	"reflect"
+	"testing"
+	"time"
 
+	"github.com/google/go-cmp/cmp"
 	"k8s.io/apimachinery/pkg/api/meta"
+	"k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/apiserver/pkg/apis/example"
 	"k8s.io/apiserver/pkg/storage"
 )
@@ -68,5 +75,99 @@ func DeepEqualSafePodSpec() example.PodSpec {
 		RestartPolicy:                 "Always",
 		TerminationGracePeriodSeconds: &grace,
 		SchedulerName:                 "default-scheduler",
+	}
+}
+
+// TestPropogateStore helps propagates store with objects, automates key generation, and returns
+// keys and stored objects.
+func TestPropogateStore(ctx context.Context, t *testing.T, store storage.Interface, obj *example.Pod) (string, *example.Pod) {
+	// Setup store with a key and grab the output for returning.
+	key := "/testkey"
+	return key, TestPropogateStoreWithKey(ctx, t, store, key, obj)
+}
+
+// TestPropogateStoreWithKey helps propagate store with objects, the given object will be stored at the specified key.
+func TestPropogateStoreWithKey(ctx context.Context, t *testing.T, store storage.Interface, key string, obj *example.Pod) *example.Pod {
+	// Setup store with the specified key and grab the output for returning.
+	err := store.Delete(ctx, key, &example.Pod{}, nil, storage.ValidateAllObjectFunc, nil)
+	if err != nil && !storage.IsNotFound(err) {
+		t.Fatalf("Cleanup failed: %v", err)
+	}
+	setOutput := &example.Pod{}
+	if err := store.Create(ctx, key, obj, setOutput, 0); err != nil {
+		t.Fatalf("Set failed: %v", err)
+	}
+	return setOutput
+}
+
+func ExpectNoDiff(t *testing.T, msg string, expected, got interface{}) {
+	t.Helper()
+	if !reflect.DeepEqual(expected, got) {
+		if diff := cmp.Diff(expected, got); diff != "" {
+			t.Errorf("%s: %s", msg, diff)
+		} else {
+			t.Errorf("%s:\nexpected: %#v\ngot: %#v", msg, expected, got)
+		}
+	}
+}
+
+const dummyPrefix = "adapter"
+
+func EncodeContinueOrDie(key string, resourceVersion int64) string {
+	token, err := storage.EncodeContinue(dummyPrefix+key, dummyPrefix, resourceVersion)
+	if err != nil {
+		panic(err)
+	}
+	return token
+}
+
+func TestCheckEventType(t *testing.T, expectEventType watch.EventType, w watch.Interface) {
+	select {
+	case res := <-w.ResultChan():
+		if res.Type != expectEventType {
+			t.Errorf("event type want=%v, get=%v", expectEventType, res.Type)
+		}
+	case <-time.After(wait.ForeverTestTimeout):
+		t.Errorf("time out after waiting %v on ResultChan", wait.ForeverTestTimeout)
+	}
+}
+
+func TestCheckResult(t *testing.T, expectEventType watch.EventType, w watch.Interface, expectObj *example.Pod) {
+	TestCheckResultFunc(t, expectEventType, w, func(object runtime.Object) error {
+		ExpectNoDiff(t, "incorrect object", expectObj, object)
+		return nil
+	})
+}
+
+func TestCheckResultFunc(t *testing.T, expectEventType watch.EventType, w watch.Interface, check func(object runtime.Object) error) {
+	select {
+	case res := <-w.ResultChan():
+		if res.Type != expectEventType {
+			t.Errorf("event type want=%v, get=%v", expectEventType, res.Type)
+			return
+		}
+		if err := check(res.Object); err != nil {
+			t.Error(err)
+		}
+	case <-time.After(wait.ForeverTestTimeout):
+		t.Errorf("time out after waiting %v on ResultChan", wait.ForeverTestTimeout)
+	}
+}
+
+func TestCheckStop(t *testing.T, w watch.Interface) {
+	select {
+	case e, ok := <-w.ResultChan():
+		if ok {
+			var obj string
+			switch e.Object.(type) {
+			case *example.Pod:
+				obj = e.Object.(*example.Pod).Name
+			case *v1.Status:
+				obj = e.Object.(*v1.Status).Message
+			}
+			t.Errorf("ResultChan should have been closed. Event: %s. Object: %s", e.Type, obj)
+		}
+	case <-time.After(wait.ForeverTestTimeout):
+		t.Errorf("time out after waiting 1s on ResultChan")
 	}
 }
